make status, late pass, until date editable on program view page add discharge to community drop down on program view page centralize bed reservation cancelation (fix discharge bug)

+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/dao/AdmissionDao.java
@@ -32,7 +32,7 @@ public interface AdmissionDao {
 
 	public List getAdmissionsInTeam(Integer programId, Integer teamId);
 
-	public Admission getTemporaryAdmission(Long demographicNo);
+	public Admission getTemporaryAdmission(Integer demographicNo);
 
 	public List search(AdmissionSearchBean searchBean);
